vuex directory 변경
