--- src/store/ations.ts
@@ -1,3 +1,0 @@
-export default {
-
-}
