+++ src/store/getters.ts
@@ -1,3 +1,8 @@
 export default {
-
+  user: (state: any) => {
+    return state.user
+  },
+  isLoading: (state: any) => {
+    return state.loading.scheduleCnt !== 0
+  }
 }
