+++ src/store/state.ts
@@ -1,3 +1,8 @@
 export default {
-
+  user: null,
+  // loading state
+  loading: {
+    status: false,
+    scheduleCnt: 0
+  }
 }
